v3/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.0

# Image processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.3.0
scikit-image>=0.22.0
scipy>=1.13.0

# OCR - install separately if needed
# paddlepaddle>=2.6.0
# paddleocr>=2.7.3
